refactor(api): migrate particpant result shared repo
