+++ api/tests/prescription/campaign-participation/integration/infrastructure/repositories/participant-results-shared-repository_test.js
@@ -1,7 +1,7 @@
-import { participantResultsSharedRepository } from '../../../../lib/infrastructure/repositories/participant-results-shared-repository.js';
-import { CampaignTypes } from '../../../../src/prescription/shared/domain/constants.js';
-import { MAX_REACHABLE_PIX_BY_COMPETENCE } from '../../../../src/shared/domain/constants.js';
-import { databaseBuilder, expect, knex, mockLearningContent } from '../../../test-helper.js';
+import { participantResultsSharedRepository } from '../../../../../../src/prescription/campaign-participation/infrastructure/repositories/participant-results-shared-repository.js';
+import { CampaignTypes } from '../../../../../../src/prescription/shared/domain/constants.js';
+import { MAX_REACHABLE_PIX_BY_COMPETENCE } from '../../../../../../src/shared/domain/constants.js';
+import { databaseBuilder, expect, knex, mockLearningContent } from '../../../../../test-helper.js';
 
 describe('Integration | Repository | Campaign Participant Result Shared Repository', function () {
   describe('#save', function () {
